Add comprehensive DMA integration and performance tests
- Introduced a new test suite for DMA integration covering various scenarios including tile data transfer, palette transfer, OAM updates, and more.
- Implemented performance tests to measure timing and efficiency of DMA transfers, including small vs large transfers, V-Blank and H-Blank constraints, and maximum concurrent DMAs.
- Added tests for DMA chaining effects, priority interrupt preemption, and IRQ overhead measurement to ensure robust functionality.
- Each test verifies memory contents post-transfer to ensure data integrity.

+++ include/dma.h
@@ -75,7 +75,7 @@ public:
     // Internal state for active transfers
     uint32_t internalSource;
     uint32_t internalDest;
-    uint16_t internalCount;
+    uint32_t internalCount;  // uint32_t to handle max count of 65536
     bool active;
 
 private:
